+++ core/index/__init__.py
@@ -1,0 +1,2 @@
+"""Index-related interfaces and future implementations."""
+